Added logic for deleting the unverified users from db and changing send to json
