+++ server/src/controllers/auth/register.js
